Renames gems to open_api-rswag-*
