+++ rswag-api/bin/rails
@@ -2,7 +2,7 @@
 # This command will automatically be run when you run "rails" with Rails 4 gems installed from the root of your application.
 
 ENGINE_ROOT = File.expand_path('../..', __FILE__)
-ENGINE_PATH = File.expand_path('../../lib/rswag/api/engine', __FILE__)
+ENGINE_PATH = File.expand_path('../../lib/open_api/rswag/api/engine', __FILE__)
 
 # Set up gems listed in the Gemfile.
 ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
