+++ rswag-specs/lib/tasks/rswag-specs_tasks.rake
@@ -8,7 +8,7 @@ namespace :rswag do
       t.pattern = 'spec/requests/**/*_spec.rb, spec/api/**/*_spec.rb, spec/integration/**/*_spec.rb'
       
       # NOTE: rspec 2.x support
-      if Rswag::Specs::RSPEC_VERSION > 2 && Rswag::Specs.config.swagger_dry_run
+      if OpenApi::Rswag::Specs::RSPEC_VERSION > 2 && OpenApi::Rswag::Specs.config.swagger_dry_run
         t.rspec_opts = [ '--format Rswag::Specs::SwaggerFormatter', '--dry-run', '--order defined' ]
       else
         t.rspec_opts = [ '--format Rswag::Specs::SwaggerFormatter', '--order defined' ]
