+++ rswag-ui/lib/generators/rswag/ui/custom/custom_generator.rb
@@ -3,7 +3,7 @@ require 'rails/generators'
 module Rswag
   module Ui
     class CustomGenerator < Rails::Generators::Base
-      source_root File.expand_path('../../../../../../lib/rswag/ui', __FILE__)
+      source_root File.expand_path('../../../../../../lib/open_api/rswag/ui', __FILE__)
 
       def add_custom_index
         copy_file('index.erb', 'app/views/rswag/ui/home/index.html.erb')
