+++ rswag-ui/lib/generators/rswag/ui/install/templates/rswag-ui.rb
@@ -1,4 +1,4 @@
-Rswag::Ui.configure do |c|
+OpenApi::Rswag::Ui.configure do |c|
 
   # List the Swagger endpoints that you want to be documented through the swagger-ui
   # The first parameter is the path (absolute or relative to the UI host) to the corresponding
