+++ rswag-ui/lib/open_api/rswag/ui/index.erb
