+++ rswag-ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "rswag-ui",
+  "name": "openapi-rswag-ui",
   "version": "1.0.0",
   "private": true,
   "dependencies": {
