+++ rswag-ui/rswag-ui.gemspec
@@ -2,11 +2,11 @@ $:.push File.expand_path("../lib", __FILE__)
 
 # Describe your gem and declare its dependencies:
 Gem::Specification.new do |s|
-  s.name        = "rswag-ui"
+  s.name        = "open_api-rswag-ui"
   s.version     = ENV['TRAVIS_TAG'] || '0.0.0'
-  s.authors     = ["Richie Morris"]
+  s.authors     = ["Richie Morris", "Jay Danielian"]
   s.email       = ["domaindrivendev@gmail.com"]
-  s.homepage    = "https://github.com/domaindrivendev/rswag"
+  s.homepage    = "https://github.com/jaydanielian/rswag"
   s.summary     = "A Rails Engine that includes swagger-ui and powers it from configured Swagger endpoints"
   s.description = "Expose beautiful API documentation, that's powered by Swagger JSON endpoints, including a UI to explore and test operations"
   s.license     = "MIT"
